[KMP] factor out GenerateFBTestClasses Gradle task
