+++ kotlin/convention-plugins/src/main/kotlin/convention.multiplatform.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+  kotlin("multiplatform")
+}
