+++ kotlin/gradle/libs.versions.toml
@@ -9,6 +9,9 @@ junit = "4.12"
 gson = "2.8.9"
 moshi-kotlin = "1.11.0"
 
+[plugins]
+kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
+
 [libraries]
 kotlin-compiler = { module = "org.jetbrains.kotlin:kotlin-compiler", version.ref = "kotlin" }
 moshi-kotlin = { module = "com.squareup.moshi:moshi-kotlin", version.ref = "moshi-kotlin" }
@@ -22,6 +25,3 @@ plugin-kotlin-gradle = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", v
 plugin-kotlinx-benchmark = { module="org.jetbrains.kotlinx:kotlinx-benchmark-plugin", version.ref="kotlinx-benchmark"}
 plugin-jmhreport = { module = "gradle.plugin.io.morethan.jmhreport:gradle-jmh-report", version="0.9.0" }
 plugin-download = { module = "de.undercouch:gradle-download-task", version = "5.3.0"}
-
-
-
